Refactor IDM functionality into separate class
