+++ src/main/java/org/folio/idmconnect/IdmClient.java
@@ -1,0 +1,9 @@
+package org.folio.idmconnect;
+
+import io.vertx.core.Future;
+import javax.ws.rs.core.Response;
+
+public interface IdmClient {
+
+  Future<Response> search(String firstName, String lastName, String dateOfBirth);
+}
