+++ src/main/java/org/folio/rest/impl/IdmConnectApi.java
@@ -1,7 +1,6 @@
 package org.folio.rest.impl;
 
 import static io.vertx.core.Future.succeededFuture;
-import static org.folio.idmconnect.Constants.MSG_IDM_URL_NOT_SET;
 import static org.folio.idmconnect.Constants.TABLE_NAME_CONTRACTS;
 
 import io.vertx.core.AsyncResult;
@@ -9,20 +8,12 @@ import io.vertx.core.CompositeFuture;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
-import io.vertx.core.buffer.Buffer;
-import io.vertx.ext.web.client.HttpRequest;
-import io.vertx.ext.web.client.HttpResponse;
-import io.vertx.ext.web.client.WebClient;
-import java.time.DateTimeException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.ResponseBuilder;
+import org.folio.idmconnect.IdmClientImpl;
 import org.folio.okapi.common.GenericCompositeFuture;
 import org.folio.rest.jaxrs.model.BulkDeleteRequest;
 import org.folio.rest.jaxrs.model.BulkDeleteResponse;
@@ -191,47 +182,6 @@ public class IdmConnectApi implements IdmConnect {
         asyncResultHandler);
   }
 
-  private String toBasicIsoDate(String dateString) {
-    try {
-      return LocalDate.parse(dateString).format(DateTimeFormatter.BASIC_ISO_DATE);
-    } catch (NullPointerException | DateTimeException e) {
-      return dateString;
-    }
-  }
-
-  private Response toResponse(HttpResponse<Buffer> bufferHttpResponse) {
-    ResponseBuilder responseBuilder =
-        Response.status(bufferHttpResponse.statusCode())
-            .header("Content-Type", bufferHttpResponse.getHeader("Content-Type"))
-            .entity(bufferHttpResponse.bodyAsString());
-    return responseBuilder.build();
-  }
-
-  private HttpRequest<Buffer> createIdmRequest(
-      WebClient webClient,
-      String idmUrl,
-      String idmToken,
-      String firstname,
-      String lastname,
-      String dateOfBirth) {
-    HttpRequest<Buffer> bufferHttpRequest = webClient.getAbs(idmUrl);
-    if (idmToken != null) {
-      bufferHttpRequest.putHeader("Authorization", idmToken);
-    }
-
-    Stream.of(
-            new String[] {"givenname", firstname},
-            new String[] {"surname", lastname},
-            new String[] {"date_of_birth", toBasicIsoDate(dateOfBirth)})
-        .forEach(
-            a -> {
-              if (a[1] != null) {
-                bufferHttpRequest.addQueryParam(a[0], a[1]);
-              }
-            });
-    return bufferHttpRequest;
-  }
-
   @Override
   public void getIdmConnectSearchidm(
       String firstname,
@@ -240,25 +190,8 @@ public class IdmConnectApi implements IdmConnect {
       Map<String, String> okapiHeaders,
       Handler<AsyncResult<Response>> asyncResultHandler,
       Context vertxContext) {
-    String idmUrl = System.getenv("IDM_URL");
-    String idmToken = System.getenv("IDM_TOKEN");
-
-    if (idmUrl == null) {
-      asyncResultHandler.handle(
-          succeededFuture(
-              GetIdmConnectSearchidmResponse.respond500WithTextPlain(MSG_IDM_URL_NOT_SET)));
-      return;
-    }
-
-    createIdmRequest(
-            WebClient.create(vertxContext.owner()),
-            idmUrl,
-            idmToken,
-            firstname,
-            lastname,
-            dateOfBirth)
-        .send()
-        .map(this::toResponse)
+    new IdmClientImpl(vertxContext)
+        .search(firstname, lastname, dateOfBirth)
         .onComplete(asyncResultHandler);
   }
 }
